requirements.txt
Django==1.10.5
djangorestframework==3.5.3
django-test-without-migrations==0.4
python-decouple==3.0
dj-database-url==0.3.0
psycopg2==2.6
coverage==4.2
django-nose==1.4.4
django-cors-headers==2.0.2